+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/RazorDiagnosticConverter.cs
@@ -41,8 +41,6 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer
             {
                 case RazorDiagnosticSeverity.Error:
                     return DiagnosticSeverity.Error;
-                case RazorDiagnosticSeverity.Warning:
-                    return DiagnosticSeverity.Warning;
                 default:
                     return DiagnosticSeverity.Information;
             }
@@ -83,7 +81,7 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer
             {
                 if (index >= sourceText.Length)
                 {
-                    // Span start index is past the end of the document. Roslyn and VSCode don't support 
+                    // Span start index is past the end of the document. Roslyn and VSCode don't support
                     // virtual positions that don't exist on the document; normalize to the last character.
                     index = sourceText.Length - 1;
                 }
