+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Semantic/SemanticTokenLegendCapability.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using OmniSharp.Extensions.LanguageServer.Protocol.Client.Capabilities;
+
+namespace Microsoft.AspNetCore.Razor.LanguageServer.Semantic
+{
+    public class SemanticTokenLegendCapability : DynamicCapability
+    {
+
+    }
+}
