+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Semantic/SemanticTokens.cs
@@ -1,0 +1,13 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System.Collections.Generic;
+
+namespace Microsoft.AspNetCore.Razor.LanguageServer.Semantic
+{
+    public class SemanticTokens
+    {
+        public string ResultId { get; }
+        public IEnumerable<uint> Data { get; set; }
+    }
+}
