+++ src/Razor/src/Microsoft.AspNetCore.Razor.VSCode.Extension/package.json
@@ -48,20 +48,8 @@
     "contributes": {
         "semanticTokenTypes": [
           {
-            "id": "minimizedTagHelperDirectiveAttribute",
-            "description": "A TagHelper Directive Attribute"
-          },
-          {
-            "id": "tagHelperDirectiveAttribute",
-            "description": "A TagHelper Directive Attribute"
-          },
-          {
-            "id": "razorTagHelperElementStartTag",
-            "description": "A Razor TagHelper Element Start Tag"
-          },
-          {
-            "id": "razorTagHelperElementEndTag",
-            "description": "A Razor TagHelper Element End Tag"
+            "id": "razorTagHelperElement",
+            "description": "A Razor TagHelper Element"
           },
           {
             "id": "razorTagHelperAttribute",
@@ -70,20 +58,8 @@
         ],
         "semanticTokenStyleDefaults": [
           {
-            "selector": "razorTagHelperElementStartTag",
-            "scope": [ "razor.taghelpers.element.starttag" ],
-            "light": {
-              "foreground": "#eb9834",
-              "fontStyle": "bold"
-            },
-            "dark": {
-              "foreground": "#eb9834",
-              "fontStyle": "bold"
-            }
-          },
-          {
-            "selector": "razorTagHelperElementEndTag",
-            "scope": [ "razor.taghelpers.element.endtag" ],
+            "selector": "razorTagHelperElement",
+            "scope": [ "razor.taghelpers.element" ],
             "light": {
               "foreground": "#eb9834",
               "fontStyle": "bold"
@@ -96,30 +72,6 @@
           {
             "selector": "razorTagHelperAttribute",
             "scope": [ "razor.taghelpers.attribute"],
-            "light": {
-              "foreground": "#eb34d9",
-              "fontStyle": "bold"
-            },
-            "dark": {
-              "foreground": "#eb34d9",
-              "fontStyle": "bold"
-            }
-          },
-          {
-            "selector": "minimizedTagHelperDirectiveAttribute",
-            "scope": [ "razor.taghelpers.attribute.directive.minimized"],
-            "light": {
-              "foreground": "#eb34d9",
-              "fontStyle": "bold"
-            },
-            "dark": {
-              "foreground": "#eb34d9",
-              "fontStyle": "bold"
-            }
-          },
-          {
-            "selector": "tagHelperDirectiveAttribute",
-            "scope": [ "razor.taghelpers.attribute.directive"],
             "light": {
               "foreground": "#eb34d9",
               "fontStyle": "bold"
