+++ src/Razor/test/VSCode.FunctionalTest/tsconfig.json
@@ -1,9 +1,9 @@
 {
-	"extends": "../../tsconfig.json",
-	"compilerOptions": {
-		"outDir": "dist"
-	},
-	"include": [
-	  "tests/**/*"
-	]
-  }
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "outDir": "dist"
+  },
+  "include": [
+    "tests/**/*"
+  ]
+}
